Re-arranged v0 schema definition
This commit moves the schema definition into a private v0 subpackage and
it also keeps the public definition mapping to the latest schema version.
The mapping of the latest schema is based on aliasing the private internal v0.

The rationale is trying to keep bits decoupled and at the same time avoiding
to repeat definitions.

Then adding a new version can be done by simply adding a new subpackage with the
new definition (which can also alias unmodified definitions from the previous
version) and extending the public package with the migration logic across versions.
