+++ pkg/manifest/api/internal/v0/testdata/full_core_release_manifest.yaml
