+++ pkg/manifest/api/internal/v0/testdata/full_solution_release_manifest.yaml
