+++ pkg/manifest/api/solution/manifest_test.go
@@ -65,7 +65,7 @@ func TestSolutionManifestSuite(t *testing.T) {
 
 var _ = Describe("ReleaseManifest", Label("release-manifest"), func() {
 	It("is parsed correctly", func() {
-		data, err := os.ReadFile(filepath.Join("..", "..", "testdata", "full_solution_release_manifest.yaml"))
+		data, err := os.ReadFile(filepath.Join("..", "internal", "v0", "testdata", "full_solution_release_manifest.yaml"))
 		Expect(err).NotTo(HaveOccurred())
 
 		rm, err := solution.Parse(data)
