+++ pkg/manifest/resolver/testdata
@@ -1,0 +1,1 @@
+../api/internal/v0/testdata
